+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.22.0] - Not released
+### Added
+- New search operators:
+  - `cliked-by:`
+  - `has:`
+  - `comments:`
+  - `likes:`
+  - `clikes:`
+  - `date:`
+  - `post-date:`
+
+  See [app/support/search/query-syntax.md](app/support/search/query-syntax.md)
+  for details.
+
 ## [2.21.0] - 2024-06-25
 ### Added
 - The V3 API is now available. The only difference from V2 is: